# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zynk API
  description: |
    End-to-end encrypted messaging platform API.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the `Authorization` header.
    Tokens are obtained via `/auth/login` or `/auth/register`.
    
    ## Rate Limiting
    - Global: 100 requests / 15 minutes per IP
    - Login: 5 attempts / 15 minutes per IP
    - Registration: 3 attempts / 15 minutes per IP
    - Messages: 60 per minute per user
    - File uploads: 20 per hour per user
    
    ## Encryption
    All message content is end-to-end encrypted using the Signal Protocol.
    The server only stores encrypted ciphertext.
  version: 1.0.0
  contact:
    name: Zynk Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    File:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        content_hash:
          type: string
        thumbnail_path:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /files/upload:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                conversation_id:
                  type: string
                  format: uuid
              required: [file]
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '429':
          description: Rate limited
  /files/{fileId}/download:
    get:
      tags: [Files]
      summary: Download a file
      operationId: downloadFile
      security:
        - BearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ======================== Keys ========================
...